rename cc namespace
