+++ include/coupledcluster/coupledcluster_mm.hpp
@@ -17,10 +17,10 @@
 #pragma once
 #include <pluginplay/pluginplay.hpp>
 
-namespace cc {
-/** @brief Loads the modules contained in the SCF module collection into the
+namespace coupledcluster {
+/** @brief Loads the modules contained in the CoupledCluster module collection into the
  *         provided ModuleManager instance.
  */
 void load_modules(pluginplay::ModuleManager& mm);
 
-} // namespace cc
+} // namespace coupledcluster
